+++ src/pages/_app.tsx
@@ -1,11 +1,15 @@
-import React from 'react'
+import  React from 'react'
+import {ProdProvider} from '../../context/prodContext/index'
 import {AppProps} from 'next/app'
 import '../styles/global.css'
+
 
 const MyApp: React.FC<AppProps> = ({ Component, pageProps }) => {
   return (
 
-     <Component {...pageProps} />
+    <ProdProvider>
+      <Component {...pageProps} />
+    </ProdProvider>
         
   )
 }
