+++ src/pages/productItem/index.js
@@ -1,7 +1,10 @@
-import React from 'react'
+import React, {useContext} from 'react'
 import Item from '../item/index'
+import { ProdContext } from '../../../context/prodContext/index'
 
-const ProductsItem = ({product, handleRemoveItemFromProd}) => {
+const ProductsItem = () => {
+  const {product} = useContext(ProdContext)
+
     return (
         <footer>
       <ul>
@@ -11,7 +14,7 @@ const ProductsItem = ({product, handleRemoveItemFromProd}) => {
          index={index}
          title={prodItem.title.substr(0,12)} 
          price={prodItem.price}
-         handleRemoveItemFromProd = {handleRemoveItemFromProd}
+     
          />
         
         )
